Fix load order bug
This fixes a bug where depending on load order, this might be loaded as
Spree::BaseHelper, causing the actual Spree::BaseHelper from solidus not
to get loaded.

Instead, I've turned this into a module prepended decorator.
